lambda_testing/scraping_lambda/pyproject.toml
[project]
name = "scraping-lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "bs4>=0.0.2",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "openai>=2.26.0",
]
